backend/modelserver/requirements.txt
flask==2.0.1
flask_cors==3.0.10
gunicorn==20.0.4
elasticsearch==7.13.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
SQLAlchemy==1.4.20
#pillow==8.3.0
tensorflow==2.5.0
tensorflow-hub==0.12.0
flasgger
Flask-Migrate==3.0.1
psycopg2-binary==2.9.1
#pandas==1.3.0
openpyxl==3.0.7
#six==1.15.0
typing_extensions

numpy
Ipython
avro-python3
apache-beam
pillow
lxml
matplotlib
Cython
contextlib2
tf_slim
six
pycocotools
lvis
scipy
pandas
google-cloud-bigquery==1.21.0
tf-models-official

celery==5.1